+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 ## 0.5.0 (in development)
 * The `pullStream` no longer requires a stream to pull, and inferes the value initially passed to the `F(...)` method.
+* Added `throttle` method on streams for limiting how often values can be pushed to a stream.
+* Added `bindPush` method on streams.